Corrected multithreadig issue and added tile
