+++ assets/Rock.material
@@ -1,0 +1,28 @@
+// generated by blender2ogre 0.8.3 on 2023-06-08 10:54:20
+material Rock {
+    receive_shadows on
+    technique {
+        pass {
+            scene_blend alpha_blend
+            cull_hardware none
+            depth_write off
+            diffuse 1.0 1.0 1.0 1.0
+            specular 0.552786 0.0 0 0 0
+
+            // additional maps - requires RTSS
+            rtshader_system {
+                lighting_stage normal_map Rock_normalMap.png
+                lighting_stage metal_roughness
+            }
+
+            // - base_color_texture
+            texture_unit {
+                texture Rock_diffuse.png
+                tex_address_mode wrap
+                colour_op modulate
+            }
+
+            // Don't know how to export: alpha_texture Rock_diffuse.png
+        }
+    }
+}
